Unit simple rework script

+++ Assets/Script/Interface/ItTarget.cs
@@ -1,10 +1,10 @@
-using BOTL.Enum;
+using BOTL.Data;
 
 public interface ItTarget
 {
     void GetDamaged(float damage);
     void Heal(float heal);
     Team GetTeam();
-    float GetTeamMultipl();
+    int GetTeamMultipl();
     float GetLife();
 }
